requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
backports.ssl-match-hostname==3.5.0.1
bintrees==2.0.7
git+https://github.com/ericsomdahl/python-bittrex.git
certifi==2017.11.5
chardet==3.0.4
coinbase==2.0.6
configparser==3.5.0
enum34==1.1.6
funcsigs==1.0.2
cbpro
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
liqui==1.0.1
mccabe==0.6.1
mock==2.0.0
pbr==3.1.1
https://github.com/s4w3d0ff/python-poloniex/archive/v0.4.6.zip
py==1.5.1
pycrypto==2.6.1
pylint==1.7.4
pytest==3.2.5
python-dateutil
requests==2.7.0
singledispatch==3.4.0.3
six==1.10.0
urllib3==1.22
websocket-client==0.40.0
wheel==0.24.0
wrapt==1.10.11
